Updated UsableItem code to fix error of immediately applying the item effect. Added a new FundsDisplayComponent to allow for the insufficient funds UI message to be handled. Changed reroll item asset
