+++ source/core/src/main/com/csse3200/game/areas/MainRoom.java
@@ -10,7 +10,8 @@ public class MainRoom extends EnemyRoom {
     @Override
     protected List<List<String>> getAnimalSpecifications() {
         return List.of(
-                List.of("Snake", "Dog"),
+                List.of("Snake"),
+//                List.of("Snake", "Dog"),
                 List.of("Bear", "Snake", "Dino"),
                 List.of("Bear", "Bear", "Minotaur"),
                 List.of("Snake", "Bat", "Minotaur"),
